Replace Hugo with homebrew generator
I was stuck on an old version of Hugo, and what I actually need is
straightforward enough that I may as well do it myself.

This also involves editing some of the posts so they still render properly;
doing it this way was easier than reverse-engineering identical markdown
parser behavior.