posts/2016-11-09-python-3d-rendering: fix figures
The hirez paths were wrong, and always have been; I didn't notice at the
time because my computer wasn't hirez! This is what I get for trying to be
clever.

I did the fixup using the generator, which means that I also had to fiddle
with a couple of other minor things along the way.